minor refactorings.
