+++ LocalizationApp/Controllers/CultureController.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Localization;
 using Microsoft.AspNetCore.Mvc;
 
-namespace LocalizationApp.Pages
+namespace LocalizationApp.Controllers
 {
     [Route("[controller]/[action]")]
     public class CultureController : Controller
@@ -15,7 +15,7 @@ namespace LocalizationApp.Pages
                     CookieRequestCultureProvider.MakeCookieValue(new RequestCulture(culture)));
             }
 
-            if (redirectionUri == null) redirectionUri = "/";
+            redirectionUri ??= "/";
 
             return LocalRedirect(redirectionUri);
