+++ LocalizationApp/Entities/AppTop.cs
@@ -1,4 +1,3 @@
-using Newtonsoft.Json;
 using System.Collections.Generic;
 
 namespace LocalizationApp.Entities
